FEATURE: Fix build gradle dependencies and more
In this commit, I fix dependencies, bump some libs versions and add Home Screen

+++ shared/src/commonMain/kotlin/com/m68476521/weather/geolocation/data/remote/models/GeoLocationDto.kt
@@ -23,7 +23,7 @@ fun GeoLocationDto.toDomain(): List<GeoLocation> =
             latitude = it.latitude,
             longitude = it.longitude,
             timeZone = it.timezone,
-            elevation = it.elevation.toDouble(),
+            elevation = it.elevation,
             countryName = it.country,
         )
     }
